Add image assignment codegen (partial)

+++ include/ImageRuntime.h
@@ -19,7 +19,7 @@ extern "C" {
         Image* load_image(const char* path);
         void save_image(Image* img, const char* path);
         void free_image(Image* img);
-        int get_image_height(Image* img);
-        int get_image_width(Image* img);
+        double get_image_height(Image* img);
+        double get_image_width(Image* img);
     }
 }
